SCRUM-3121 moved error components
