+++ src/main/cliapp/src/containers/allelesPage/GermlineTransmissionStatusDialog.js
@@ -7,7 +7,7 @@ import { Toast } from 'primereact/toast';
 import { ColumnGroup } from 'primereact/columngroup';
 import { Row } from 'primereact/row';
 import { TrueFalseDropdown } from '../../components/TrueFalseDropDownSelector';
-import { DialogErrorMessageComponent } from '../../components/DialogErrorMessageComponent';
+import { DialogErrorMessageComponent } from '../../components/Error/DialogErrorMessageComponent';
 import { EllipsisTableCell } from '../../components/EllipsisTableCell';
 import { ControlledVocabularyDropdown } from '../../components/ControlledVocabularySelector';
 import { useControlledVocabularyService } from '../../service/useControlledVocabularyService';
