+++ src/main/cliapp/src/containers/vocabularyTermSetPage/VocabularyTermSetTable.js
@@ -3,7 +3,7 @@ import { useMutation } from 'react-query';
 import { Toast } from 'primereact/toast';
 import { SearchService } from '../../service/SearchService';
 import { Messages } from 'primereact/messages';
-import { ErrorMessageComponent } from "../../components/ErrorMessageComponent";
+import { ErrorMessageComponent } from "../../components/Error/ErrorMessageComponent";
 import { EllipsisTableCell } from "../../components/EllipsisTableCell";
 import { ListTableCell } from "../../components/ListTableCell";
 import { Button } from 'primereact/button';
